ingestion/requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
flask==3.0.0
